requirements.txt
django==3.0.6
mysqlclient==1.4.4
Pillow==7.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
sorl-thumbnail==12.6.2